refactor(public-witnessing): split the location form and make day rows keyboard-operable
